+++ R/LearnerClustMiniBatchKMeans.R
@@ -48,7 +48,8 @@ LearnerClustMiniBatchKMeans = R6Class(
         CENTROIDS = p_uty(default = NULL, tags = "train"),
         tol = p_dbl(0, default = 1e-04, tags = "train"),
         tol_optimal_init = p_dbl(0, default = 0.3, tags = "train"),
-        seed = p_int(default = 1L, tags = "train")
+        seed = p_int(default = 1L, tags = "train"),
+        threads = p_int(1L, default = 1L, tags = c("predict", "threads"))
       )
 
       param_set$set_values(clusters = 2L)
@@ -77,17 +78,22 @@ LearnerClustMiniBatchKMeans = R6Class(
       data = task$data()
       m = invoke(ClusterR::MiniBatchKmeans, data = data, .args = pv)
       if (self$save_assignments) {
-        self$assignments = as.integer(invoke(ClusterR::predict_KMeans, data = data, CENTROIDS = m$centroids))
+        pv_predict = self$param_set$get_values(tags = "predict")
+        self$assignments = as.integer(
+          invoke(ClusterR::predict_KMeans, data = data, CENTROIDS = m$centroids, .args = pv_predict)
+        )
       }
       m
     },
 
     .predict = function(task) {
+      pv = self$param_set$get_values(tags = "predict")
       data = ordered_features(task, self)
-      partition = as.integer(invoke(ClusterR::predict_KMeans, data = data, CENTROIDS = self$model$centroids))
+      centroids = self$model$centroids
+      partition = as.integer(invoke(ClusterR::predict_KMeans, data = data, CENTROIDS = centroids, .args = pv))
       prob = NULL
       if (self$predict_type == "prob") {
-        prob = invoke(ClusterR::predict_KMeans, data = data, CENTROIDS = self$model$centroids, fuzzy = TRUE)
+        prob = invoke(ClusterR::predict_KMeans, data = data, CENTROIDS = centroids, fuzzy = TRUE, .args = pv)
         colnames(prob) = seq_col(prob)
       }
       list(partition = partition, prob = prob)